requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.12.1
Pillow==7.2.0
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
